pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-mcp-adapters>=0.1.9",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
    "a2a-sdk>=0.3.3",
    "requests>=2.32.0",
    "requests-cache>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "urllib3>=2.0.0",
    "pandas>=2.2.0",
    "loguru>=0.7.0",
    "mcp[cli]>=1.10.1",
    "langgraph>=0.6.6",
]
